breaking mod
this breaks xunit. further work needed...
